==> R/2.TFBS_rates.Rmd <==
---
title: "R Notebook"
output: html_notebook
editor_options: 
  chunk_output_type: console
---



```{r}
library(reshape2)
library(tidyverse)
library(ggplot2); theme_set(theme_bw())

dataset <- read.csv("../data/all_data_bcd_2019_10_01_after_QC.csv")
```

## Checking it out
```{r}
head(dataset)

## Set data types
str(dataset)
dataset$raw_seq <- as.character(dataset$raw_seq)
dataset$enhancer_func <- as.factor(dataset$enhancer_func)
```

### Looking at postion specific across the whole dataset

Code taken from Zoe's work.
```{r}
#group by lowercase factors and sum over n
grouped_seq <- dataset %>%
  group_by(raw_seq) %>% 
  mutate(n= n()) %>% 
  distinct(raw_seq, .keep_all=TRUE) %>%
  arrange(desc(n))
  
#create a column containing lists of bases
ordered_seq$split_seq = strsplit(ordered_seq$raw_seq, "")
ordered_seq

#assign a column to each base based on position
pos_seq <- ordered_seq %>% 
  separate(split_seq, c("pos0", "pos1", "pos2", "pos3", "pos4", "pos5", "pos6")) %>%
  select(-c("pos0"))

pos_seq_reshaped <- melt(pos_seq, measure.vars = c("pos1", "pos2", "pos3", "pos4", "pos5", "pos6"))

pos_seq_groups <- pos_seq_reshaped %>% group_by(enhancer_func, variable) %>%
             mutate(group_size = n()) %>%
             group_by(variable, value, enhancer_func) %>%
             summarise(perc = n()/max(group_size))

colnames(pos_seq_groups)[colnames(pos_seq_groups)=="variable"] <- "positions"
colnames(pos_seq_groups)[colnames(pos_seq_groups)=="value"] <- "nucleotides"
colnames(pos_seq_groups)[colnames(pos_seq_groups)=="perc"] <- "percentage"

pos_seq_groups

ggplot(data=pos_seq_groups, aes(fill=enhancer_func, y=percentage, x=nucleotides)) + 
  geom_bar(stat="identity", position="dodge") + 
  scale_y_continuous(labels = scales::percent) + 
  scale_fill_grey() +
  facet_grid(~positions) 
```

## Now looking at each region more carefully 

In the end, we will have to look at each region as it's own instance.  So, lets do some exploritory work looking at each region. Maybe by looking at nucleotide diversity at each position? 

So essentially, we would be doing group_by, and calculating allele frequency at each postion, sum and multiple by 1/6.  Then see how each position differs depending on nucleotide fucntnion. This is where we would need some better controls. 

One thing niharika coudl do is create these control experiments. 

```{r}
#create a column containing lists of bases
dataset$split_seq = strsplit(dataset$raw_seq, "")

#assign a column to each base based on position
datset <- dataset %>% 
  separate(split_seq, c("pos0", "pos1", "pos2", "pos3", "pos4", "pos5", "pos6")) %>%
  select(-c("pos0"))

## dataset_long <- melt(pos_seq, measure.vars = c("pos1", "pos2", "pos3", "pos4", "pos5", "pos6"))

dataset %>%
  group_by(region, align_position, strand) %>%
```
